Rename data generator DOM
